# 🚀 Performance Optimization: Comprehensive React Portfolio Enhancement