--- a/package.json
+++ b/package.json
@@ -6,36 +6,78 @@
     "start": "webpack serve --mode development --open",
     "build": "webpack --mode production",
     "build:gh-pages": "cross-env GITHUB_PAGES=true webpack --mode production",
+    "build:analyze": "cross-env ANALYZE=true npm run build",
+    "build:stats": "webpack --mode production --json > webpack-stats.json",
+    "preview": "http-server dist -p 8080 -c-1",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "lint:fix": "eslint . --ext js,jsx --fix",
     "test": "jest",
-    "deploy": "npm run build:gh-pages && node deploy-gh-pages.cjs"
+    "test:coverage": "jest --coverage",
+    "deploy": "npm run build:gh-pages && node deploy-gh-pages.cjs",
+    "size-limit": "size-limit",
+    "typecheck": "tsc --noEmit"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "core-js": "^3.33.0"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@testing-library/jest-dom": "^6.9.0",
     "@testing-library/react": "^16.3.0",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
+    "autoprefixer": "^10.4.16",
     "babel-loader": "^9.0.0",
     "clean-webpack-plugin": "^4.0.0",
+    "compression-webpack-plugin": "^10.0.0",
     "cross-env": "^7.0.3",
     "css-loader": "^6.0.0",
+    "css-minimizer-webpack-plugin": "^5.0.1",
+    "cssnano": "^6.0.1",
     "eslint": "^8.55.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "html-webpack-plugin": "^5.0.0",
     "http-server": "^14.1.1",
+    "imagemin-webpack-plugin": "^2.4.2",
     "jest": "^30.2.0",
     "mini-css-extract-plugin": "^2.0.0",
+    "postcss": "^8.4.31",
+    "postcss-loader": "^7.3.3",
+    "size-limit": "^9.0.0",
     "style-loader": "^4.0.0",
+    "terser-webpack-plugin": "^5.3.9",
+    "typescript": "^5.2.2",
     "webpack": "^5.0.0",
+    "webpack-bundle-analyzer": "^4.9.1",
     "webpack-cli": "^5.0.0",
     "webpack-dev-server": "^4.0.0"
-  }
+  },
+  "size-limit": [
+    {
+      "path": "dist/assets/js/*.js",
+      "limit": "250 KB"
+    },
+    {
+      "path": "dist/assets/css/*.css",
+      "limit": "50 KB"
+    }
+  ]
 }