--- a/webpack.config.cjs
+++ b/webpack.config.cjs
@@ -5,72 +5,194 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
 module.exports = (env, argv) => {
   const isProduction = argv.mode === 'production';
+  const isDevelopment = !isProduction;
   
   return {
-  entry: './src/main.jsx',
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-    filename: isProduction ? 'assets/[name].[contenthash].js' : 'assets/[name].js',
-    publicPath: process.env.GITHUB_PAGES === 'true' ? '/kumar-ravi-portfolio/' : '/',
-    clean: true,
-  },
-  resolve: {
-    extensions: ['.js', '.jsx'],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(js|jsx)$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: [
-              '@babel/preset-env',
-              ['@babel/preset-react', { runtime: 'automatic' }]
-            ],
+    entry: {
+      main: './src/main.jsx',
+    },
+    output: {
+      path: path.resolve(__dirname, 'dist'),
+      filename: isProduction 
+        ? 'assets/js/[name].[contenthash:8].js' 
+        : 'assets/js/[name].js',
+      chunkFilename: isProduction 
+        ? 'assets/js/[name].[contenthash:8].chunk.js'
+        : 'assets/js/[name].chunk.js',
+      publicPath: process.env.GITHUB_PAGES === 'true' ? '/kumar-ravi-portfolio/' : '/',
+      clean: true,
+      assetModuleFilename: 'assets/media/[name].[contenthash:8][ext]',
+    },
+    resolve: {
+      extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      alias: {
+        '@': path.resolve(__dirname, 'src'),
+        '@components': path.resolve(__dirname, 'src/components'),
+        '@utils': path.resolve(__dirname, 'src/utils'),
+        '@hooks': path.resolve(__dirname, 'src/hooks'),
+        '@constants': path.resolve(__dirname, 'src/constants'),
+        '@contexts': path.resolve(__dirname, 'src/contexts'),
+      },
+    },
+    module: {
+      rules: [
+        {
+          test: /\.(js|jsx|ts|tsx)$/,
+          exclude: /node_modules/,
+          use: {
+            loader: 'babel-loader',
+            options: {
+              presets: [
+                ['@babel/preset-env', {
+                  useBuiltIns: 'usage',
+                  corejs: 3,
+                  modules: false,
+                  targets: {
+                    browsers: ['> 1%', 'last 2 versions', 'not ie <= 11']
+                  }
+                }],
+                ['@babel/preset-react', { runtime: 'automatic' }]
+              ],
+              plugins: [
+                ['@babel/plugin-syntax-dynamic-import'],
+              ],
+              cacheDirectory: true,
+              cacheCompression: false,
+            },
+          },
+        },
+        {
+          test: /\.css$/,
+          use: [
+            isProduction ? {
+              loader: MiniCssExtractPlugin.loader,
+              options: {
+                publicPath: '../../',
+              },
+            } : 'style-loader',
+            {
+              loader: 'css-loader',
+              options: {
+                importLoaders: 1,
+                sourceMap: isDevelopment,
+              },
+            },
+          ],
+        },
+        {
+          test: /\.(png|svg|jpg|jpeg|gif|webp)$/i,
+          type: 'asset',
+          parser: {
+            dataUrlCondition: {
+              maxSize: 8 * 1024, // 8kb - inline small images
+            },
+          },
+          generator: {
+            filename: 'assets/images/[name].[contenthash:8][ext]',
+          },
+        },
+        {
+          test: /\.(woff|woff2|eot|ttf|otf)$/i,
+          type: 'asset/resource',
+          generator: {
+            filename: 'assets/fonts/[name].[contenthash:8][ext]',
+          },
+        },
+        {
+          test: /\.(pdf|doc|docx)$/i,
+          type: 'asset/resource',
+          generator: {
+            filename: 'assets/documents/[name].[contenthash:8][ext]',
+          },
+        },
+      ],
+    },
+    plugins: [
+      new CleanWebpackPlugin(),
+      new HtmlWebpackPlugin({
+        template: './index.html',
+        filename: 'index.html',
+        inject: 'body',
+        scriptLoading: 'defer',
+        minify: isProduction && {
+          removeComments: true,
+          collapseWhitespace: true,
+          removeRedundantAttributes: true,
+          useShortDoctype: true,
+          removeEmptyAttributes: true,
+          removeStyleLinkTypeAttributes: true,
+          keepClosingSlash: true,
+          minifyJS: true,
+          minifyCSS: true,
+          minifyURLs: true,
+        },
+      }),
+      isProduction && new MiniCssExtractPlugin({
+        filename: 'assets/css/[name].[contenthash:8].css',
+        chunkFilename: 'assets/css/[name].[contenthash:8].chunk.css',
+      }),
+    ].filter(Boolean),
+    optimization: {
+      minimize: isProduction,
+      splitChunks: {
+        chunks: 'all',
+        cacheGroups: {
+          default: {
+            minChunks: 2,
+            priority: -20,
+            reuseExistingChunk: true,
+          },
+          vendor: {
+            test: /[\/]node_modules[\/]/,
+            name: 'vendors',
+            priority: -10,
+            chunks: 'all',
+            enforce: true,
+          },
+          react: {
+            test: /[\/]node_modules[\/](react|react-dom)[\/]/,
+            name: 'react-vendors',
+            priority: 20,
+            chunks: 'all',
+            enforce: true,
+          },
+          common: {
+            name: 'common',
+            minChunks: 2,
+            priority: 5,
+            chunks: 'all',
+            enforce: true,
           },
         },
       },
-      {
-        test: /\.css$/,
-        use: [
-          isProduction ? MiniCssExtractPlugin.loader : 'style-loader',
-          'css-loader'
-        ],
+      runtimeChunk: {
+        name: 'runtime',
       },
-      {
-        test: /\.(png|svg|jpg|jpeg|gif|pdf)$/i,
-        type: 'asset/resource',
-        generator: {
-          filename: 'assets/[name].[contenthash][ext]',
-        },
+      usedExports: true,
+      sideEffects: false,
+      moduleIds: 'deterministic',
+    },
+    performance: {
+      hints: isProduction ? 'warning' : false,
+      maxEntrypointSize: 512000, // 500kb
+      maxAssetSize: 512000, // 500kb
+    },
+    devServer: {
+      static: {
+        directory: path.join(__dirname, 'public'),
       },
-    ],
-  },
-  plugins: [
-    new CleanWebpackPlugin(),
-    new HtmlWebpackPlugin({
-      template: './index.html',
-      filename: 'index.html',
-    }),
-    ...(isProduction ? [
-      new MiniCssExtractPlugin({
-        filename: 'assets/[name].[contenthash].css',
-      })
-    ] : []),
-  ],
-  optimization: {
-    splitChunks: {
-      chunks: 'all',
-      cacheGroups: {
-        vendor: {
-          test: /[\\/]node_modules[\\/]/,
-          name: 'vendors',
-          chunks: 'all',
+      compress: true,
+      port: 3000,
+      open: true,
+      hot: true,
+      historyApiFallback: true,
+      client: {
+        overlay: {
+          errors: true,
+          warnings: false,
         },
       },
     },
-  },
+    devtool: isProduction ? 'source-map' : 'eval-cheap-module-source-map',
   };
 };